chore: MainDIContainer 삭제
